common classes moved to communication folder
