+++ client/communication/ClientCommunicator.h
@@ -2,13 +2,13 @@
 #define COMMUNICATOR_H
 
 #include <string>
-#include <common/StageInfo.h>
+#include <common/communication/StageInfo.h>
 
-#include "common/Receiver.h"
-#include "common/Packet.h"
-#include "common/Socket.h"
-#include "common/QuitProtected.h"
-#include "common/Sender.h"
+#include "common/communication/Receiver.h"
+#include "common/communication/Packet.h"
+#include "common/communication/Socket.h"
+#include "common/communication/QuitProtected.h"
+#include "common/communication/Sender.h"
 
 class ClientReceiver : public Receiver {
     private:
