+++ common/communication/QuitProtected.cpp
