+++ common/communication/Receiver.cpp
