+++ common/communication/StageInfo.cpp
