+++ common/communication/StageInfo.h
