+++ server/model/Game.h
@@ -5,7 +5,7 @@
 #include <vector>
 #include <common/Thread.h>
 
-#include "common/StageInfo.h"
+#include "common/communication/StageInfo.h"
 #include "Player.h"
 #include "Map.h"
